rearrange test structure
